Adiciona endpoint /absent para marcar um agendamento como ausente.
